pyproject.toml
[tool.poetry]
name = "data-migration"
version = "0.1.0"
description = ""
authors = ["obenchekro <othmane.benchekroun@supinfo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyodbc = "^5.0.1"
pyspark = "^3.5.0"
kafka-python = "^2.0.2"
azure-storage-blob = "^12.19.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
